rgw: move rgw_raw_obj to rgw_basic_types.{h,cc}
